1. Добавил простую пагинацию 2. Добавил общее количество элементов параметры методов репозитория 3. Добавил хелпер для общего количества и з/п

+++ VacancyStore.DataAccess.Remote/HHApi/Responses/GetVacanciesResponse.cs
@@ -14,7 +14,7 @@ namespace VacancyStore.DataAccess.Remote.Responses
 
         public string Arguments { get; set; }
 
-        public int Found { get; set; }
+        public long Found { get; set; }
 
         [JsonProperty("alternate_url")]
         public string AlternateUrl { get; set; }
